requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
jinja2==3.1.4
python-multipart==0.0.9
sqlalchemy==2.0.32
psycopg[binary]==3.2.1
pgvector==0.2.5
openai==1.42.0
langchain
langchain-openai
cohere==5.6.2
langgraph
beautifulsoup4==4.12.3
lxml==5.3.0
pydantic==2.8.2
python-dotenv==1.0.1
selenium==4.23.1
websockets==12.0
watchfiles==0.22.0
